requirements.txt
fredapi
pandas
matplotlib
reportlab
requests
beautifulsoup4
anthropic
python-dotenv
tavily-python
